+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@
   * [`ProxyJump` support with paramiko ssh backend](https://github.com/Xpra-org/xpra/commit/2167c90e98555d4d62d62924fdb868eba5f96888)
   * [support remote UNC named-pipe URLs](https://github.com/Xpra-org/xpra/issues/1454)
   * [much improved RFB support](https://github.com/Xpra-org/xpra/issues/4933)
+  * [prefixed packets](https://github.com/Xpra-org/xpra/issues/4972)
   * [document current protocol](https://github.com/Xpra-org/xpra/issues/4999)
 * 🛡️ Security:
   * [`seccomp` hardening of image parsing](https://github.com/Xpra-org/xpra/issues/622)